+++ src/types.ts
@@ -21,6 +21,7 @@ export interface Options {
 	childLanguages?: boolean
 	quick?: boolean
 	offline?: boolean
+	calculateLines?: boolean
 	checkIgnored?: boolean
 	checkDetected?: boolean
 	checkAttributes?: boolean
@@ -33,6 +34,11 @@ export interface Results {
 	files: {
 		count: Integer
 		bytes: Bytes
+		lines: {
+			total: Integer
+			content: Integer
+			code: Integer
+		}
 		/** Note: Results use slashes as delimiters even on Windows. */
 		results: Record<FilePath, LanguageResult>
 		alternatives: Record<FilePath, LanguageResult[]>
@@ -40,8 +46,18 @@ export interface Results {
 	languages: {
 		count: Integer
 		bytes: Bytes
+		lines: {
+			total: Integer
+			content: Integer
+			code: Integer
+		}
 		results: Record<Language, {
 			bytes: Bytes
+			lines: {
+				total: Integer
+				content: Integer
+				code: Integer
+			}
 			type: Category
 			parent?: Language
 			color?: `#${string}`
@@ -50,6 +66,11 @@ export interface Results {
 	unknown: {
 		count: Integer
 		bytes: Bytes
+		lines: {
+			total: Integer
+			content: Integer
+			code: Integer
+		}
 		extensions: Record<string, Bytes>
 		filenames: Record<string, Bytes>
 	}